refactor move App class to logs package
