+++ app/src/main/java/com/parishod/watomatic/model/logs/App.kt
@@ -1,4 +1,4 @@
-package com.parishod.watomatic.model
+package com.parishod.watomatic.model.logs
 
 data class App(
         val name:String,
